fix compile for 64bit
svn path=/trunk/kdemultimedia/libkcddb/; revision=255039

+++ cddb.cpp
@@ -149,7 +149,7 @@ namespace KCDDB
 
     QByteArray buf( maxRead );
     Q_LONG bytesRead = socket_.readLine( buf.data(), maxRead - 1 );
-    buf[ bytesRead ] = '\0';
+    buf[ (int)bytesRead ] = '\0';
 
     return QString( buf );
   }
